refactor: Factor out duplicate local functions
